Make groups struct ser/de
