+++ libsignal-service/src/groups_v2/mod.rs
@@ -1,5 +1,6 @@
 //! Everything needed to support [Signal Groups v2](https://signal.org/blog/new-groups/)
 mod manager;
+mod model;
 mod operations;
 pub mod utils;
 
@@ -7,4 +8,5 @@ pub use manager::{
     decrypt_group, CredentialsCache, CredentialsCacheError, GroupsManager,
     InMemoryCredentialsCache,
 };
-pub use operations::{Group, GroupChange, GroupChanges, GroupDecryptionError};
+pub use model::{Group, GroupChange, GroupChanges};
+pub use operations::GroupDecryptionError;
